pyproject.toml
[build-system]
requires = ["setuptools>=65.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "dj-core-utils"
version = "0.1.0"
description = "utilidades para Django monolito y/o microservicios"
readme = "README.md"
authors = [
    { name = "Joaquin Lugo Zavala", email = "joaquinlugo123@gmail.com" }
]
license = {text = ""}
requires-python = ">=3.10"
dependencies = [
    "django>=5.2",
    "requests>=2.31",
    "python-dotenv>=1.0",
    "djangorestframework>=3.14",
    "djangorestframework-simplejwt>=5.3.0",
    "django-cors-headers>=4.3.1",
    "pyjwt>=2.8.0",
    "django-currentuser>=0.9.0",
    "python-json-logger>=3.3.0",
    "markdown>=3.8",
    "django-filter>=25.1",
    "python-jose>=3.0.4",
    "django-prometheus>=2.3.1",
    "prometheus-client>=0.21.1",
    "psycopg2-binary>=2.9.10",
]

[project.optional-dependencies]
fastapi = [  # Grupo opcional
    "fastapi>=0.95.0",
    "uvicorn>=0.21.0",
    "python-multipart>=0.0.5",
    "sqlalchemy>=2.0.0",
    "pydantic>=2.0.0"
]
dev = [
    "pytest>=7.0",
    "black>=23.0",
    "httpx>=0.23.0"  # Para testing de APIs
]

[tool.setuptools.packages.find]
where = ["src"]
include = ["dj_core_utils*"]

[tool.mypy]
strict = true
ignore_missing_imports = true